Extract createQueryClient to a module
